requirements.txt
pydantic>=2.0.0
faker>=20.0.0
asteval>=0.9.31
google-cloud-storage>=2.10.0
boto3>=1.28.0
PyYAML>=6.0.1
